refactor : 컨버터 이름 변경 및 사용 위치 새로운 이름으로 갱신
