+++ src/main/java/prgrms/marco/be02marbox/domain/theater/service/TheaterService.java
@@ -3,7 +3,7 @@ package prgrms.marco.be02marbox.domain.theater.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import prgrms.marco.be02marbox.domain.converter.TheaterConverter;
+import prgrms.marco.be02marbox.domain.theater.service.utils.TheaterConverter;
 import prgrms.marco.be02marbox.domain.theater.Theater;
 import prgrms.marco.be02marbox.domain.theater.dto.RequestCreateTheater;
 import prgrms.marco.be02marbox.domain.theater.repository.TheaterRepository;
@@ -21,7 +21,7 @@ public class TheaterService {
 
 	@Transactional
 	public Long createTheater(RequestCreateTheater request) {
-		Theater newTheater = theaterConverter.getTheater(request);
+		Theater newTheater = theaterConverter.convertFromRequestCreateTheaterToTheater(request);
 		Theater savedTheater = theaterRepository.save(newTheater);
 		return savedTheater.getId();
 	}
